% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eonet.R
\name{earth_event}
\alias{earth_event}
\title{Calls EONET webservice}
\usage{
earth_event(status = "all", sources = "all", category_id = "all",
  limit = 10, days = 20, LimitType = "limit", TrySimplify = TRUE)
}
\arguments{
\item{status}{Accepts 1 or 0 (open or closed). Defaults to "all", which includes both.}

\item{sources}{Accepts character id strings from EONET sources (see \code{eonet_sources})}

\item{category_id}{bla}

\item{limit}{vla}

\item{days}{jks}

\item{LimitType}{lvo}

\item{TrySimplify}{lasn}
}
\description{
Calls NASA's Earth Observatory Natural Event Tracker (EONET) webservice and returns a data.frame with individual event or events.
}
\details{

}

