% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eonet.R
\name{eonet_categories}
\alias{eonet_categories}
\title{Calls EONET category webservice}
\usage{
eonet_categories(limit = 10)
}
\description{
Calls NASA's Earth Imagery API and returns data.frame with information on time
}

