% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eonet.R
\name{eonet_sources}
\alias{eonet_sources}
\title{Calls EONET sources webservice}
\usage{
eonet_sources()
}
\value{
Returns \code{data.frame} with 4 columns:
\item{id}{Unique id (can be used to filter \code{earth_event})}
\item{title}{Title of source}
\item{source}{Official source URL}
\item{link}{Direct json link (the result is equal to filtering all
            \code{earth_event} with source)}
}
\description{
Calls NASA's EONET Webservice and returns all sources available.
}
\examples{
\dontrun{
sources <- eonet_sources()
}
}

