% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{trace_sat}
\alias{trace_sat}
\title{Trace a satellite route}
\usage{
trace_sat(key, lon1, lon2, lat1, lat2, date1, date2, cloud_score = TRUE,
  plot = FALSE, meta_only = FALSE)
}
\arguments{
\item{key}{Key for API authentication.}
}
\description{
Calls NASA's Earth Imagery Assets API and returns data.frame with information on time and location between two dates to "trace" a satellites path.
}

