% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eonet.R
\name{eonet_sources}
\alias{eonet_sources}
\title{Calls EONET sources webservice}
\usage{
eonet_sources()
}
\description{
Calls NASA's EONET Webservice and returns all sources available.
}
\details{
Returns data.frame with 4 columns
 @field id Unique id (can be used to filter \code{earth_event})
 @field title Title of source
 @field source Official source URL
 @field link Direct json link (the result is equal to filtering all \code{earth_event} with source)
}
\examples{
\dontrun{
sources <- eonet_sources()
}
}

