% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eonet.R
\name{eonet_sources}
\alias{eonet_sources}
\title{Calls EONET sources webservice}
\usage{
eonet_sources(limit = 10)
}
\description{
Calls NASA's Earth Imagery API and returns data.frame with information on time
}

