% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imagery.R
\name{plot_earth_image}
\alias{plot_earth_image}
\title{Plots the image to device}
\usage{
plot_earth_image(image_png)
}
\arguments{
\item{image_png}{image downloaded using earth_image.}
}
\description{
To avoid S4 Classes and methods, this small wrapper simply plots an image from NASA.
If the purpose is to this interactively on one image, set the parameter plot = TRUE in \code{earth_image}
}
\examples{
\dontrun{
key <- "123key"
img <- earth_image(key, -100.31008, 25.66779, "2016-01-01")
plot_earth_image(img$image_png)
}
}
\seealso{
earth_image
}

